src/chorus_bridge/services/conductor_client.py
from typing import List, Any
from chorus.conductor.conductor import Conductor
from chorus.conductor.models import Event, PostAnnounce, ModerationEvent, UserRegistration, DayProof, MembershipChange, APExportNotice

class ConductorClient:
    """Client for interacting with the local Conductor instance."""

    def __init__(self, conductor_instance: Conductor):
        self.conductor = conductor_instance

    async def submit_events(self, events: List[Event]) -> None:
        """Submits a list of events to the Conductor for consensus.

        Args:
            events: A list of Event objects to be submitted.
        """
        print(f"ConductorClient: Submitting {len(events)} events to Conductor.")
        await self.conductor.propose_batch(events)

    async def get_committed_events(self) -> List[Any]:
        """Placeholder for retrieving committed events from Conductor.

        In a real implementation, this would involve a mechanism for Conductor
        to notify the Bridge of new committed blocks, or the Bridge polling Conductor.
        """
        print("ConductorClient: Retrieving committed events from Conductor (simulated).")
        # For now, we'll just return a dummy list of events
        return []
