pyproject.toml
[project]
name = "chorus-bridge"
version = "0.1.0"
description = ""
authors = [
    {name = "Hailey Welsh",email = "github@haileywelsh.me"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "httpx>=0.27.0",
    "PyNaCl>=1.5.0",
    "blake3>=0.3.4",
    "sqlalchemy>=2.0.27",
    "psycopg[binary]>=3.2.1",
    "protobuf>=4.25.3",
    "grpcio-tools>=1.62.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0"
]

[tool.poetry]
packages = [{include = "chorus_bridge", from = "src"}]

[tool.poetry.dependencies]
conductor = { path = "../conductor", develop = true }

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)"
]
